pyproject.toml
[tool.poetry]
name = "korean-blog-extractor"
version = "0.8.0"
description = ""
authors = ["Michael Kang <michael3dk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
pyyaml = "^6.0.1"
feedparser = "^6.0.10"
httpx = "^0.28.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
pytest-mock = "^3.11.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
